refactor(console): move and rename global angularjs module
now the management folder is only for pages related to the selected environment
